[easy] Remove script to run in-place experiment builder.
Previous versions of the in-place experiment builder required monkey patching the code to work.  Since this is no longer necessary, we do not have to isolate the in-place building to its own process.  Remove the script and directly build the experiment.

Test plan: `pytest -n4 -v starfish/core/experiment/builder/test/`
