+++ starfish/core/experiment/builder/test/structured_formatter/test_single_fov.py
@@ -11,7 +11,6 @@ from starfish.core.experiment.experiment import Experiment, FieldOfView
 from starfish.core.imagestack.test.imagestack_test_utils import verify_physical_coordinates
 from starfish.core.types import Axes, Coordinates, CoordinateValue
 from .util import (
-    format_data,
     render_coordinates_to_rows,
     UniqueTiles,
     write_coordinates_csv,
@@ -23,6 +22,7 @@ from ... import (
     TileFetcher,
     TileIdentifier,
 )
+from ...structured_formatter import format_structured_dataset
 
 
 @pytest.mark.parametrize(
@@ -94,7 +94,7 @@ def test_single_aligned_regular_fov(
     if tile_format == ImageFormat.TIFF:
         time.sleep(1)
 
-    format_data(
+    format_structured_dataset(
         os.fspath(tmpdir_path),
         os.fspath(coordinates_csv_path),
         os.fspath(tmpdir_path),
@@ -185,7 +185,7 @@ def test_single_ragged_fov(
     rows = render_coordinates_to_rows(tile_coordinates)
     write_coordinates_csv(coordinates_csv_path, rows, True)
 
-    format_data(
+    format_structured_dataset(
         os.fspath(tmpdir_path),
         os.fspath(coordinates_csv_path),
         os.fspath(tmpdir_path),
